Large changes to package, especially the cli. Now using click instead of argparse as this massively simplifies the cli. Edits to many scripts to remove all classes and place them into appropriate modules. Pipeline slightly edited to enable it to function with these changes but more to be done
