--- ccanalyser/__init__.py
@@ -1,3 +1,0 @@
-# import ccanalyser.ccanalysis
-# import ccanalyser.utils
-# import ccanalyser.pipeline
