--- ccanalyser/ccanalysis/__init__.py
@@ -1,3 +1,0 @@
-from ccanalyser.ccanalysis import annotate_slices
-from ccanalyser.ccanalysis import ccanalysis
-from ccanalyser.ccanalysis import store_interactions
