+++ ccanalyser/tools/filter.py
@@ -1,165 +1,7 @@
-import os
-import time
-from datetime import timedelta
-from functools import wraps
-
-import numpy as np
 import pandas as pd
-import pysam
+import os
+import numpy as np
 
-
-def get_timing(task_name=None):
-    """Decorator:
-    Gets the time taken by the wrapped function
-    """
-
-    def wrapper(f):
-        @wraps(f)
-        def wrapped(*args, **kwargs):
-            time_start = time.perf_counter()
-            result = f(*args, **kwargs)
-            time_end = time.perf_counter()
-
-            time_taken = timedelta(seconds=(time_end - time_start))
-            print(f"Completed {task_name} in {time_taken} (hh:mm:ss.ms)")
-            return result
-
-        return wrapped
-
-    return wrapper
-
-
-def parse_alignment(aln):
-    """Parses reads from a bam file into a list.
-
-    Args:
-      aln: pysam.AlignmentFile
-    Returns:
-      List containing:
-      -read name
-      -parent reads
-      -flashed status
-      -slice number
-      -mapped status
-      -multimapping status
-      -chromosome number (e.g. chr10)
-      -start (e.g. 1000)
-      -end (e.g. 2000)
-      -coords e.g. (chr10:1000-2000)
-    """
-
-    slice_name = aln.query_name
-    parent_read, pe, slice_number, uid = slice_name.split("|")
-    ref_name = aln.reference_name
-    ref_start = aln.reference_start
-    ref_end = aln.reference_end
-    # Check if read mapped
-    if aln.is_unmapped:
-        mapped = 0
-        multimapped = 0
-        ref_name = "unmapped"
-        ref_start = ""
-        ref_end = ""
-        coords = ""
-    else:
-        mapped = 1
-        coords = f"{ref_name}:{ref_start}-{ref_end}"
-        # Check if multimapped
-        if aln.is_secondary:
-            multimapped = 1
-        else:
-            multimapped = 0
-    return [
-        slice_name,
-        parent_read,
-        pe,
-        slice_number,
-        uid,
-        mapped,
-        multimapped,
-        ref_name,
-        ref_start,
-        ref_end,
-        coords,
-    ]
-
-
-@get_timing(task_name="processing BAM file")
-def parse_bam(bam):
-    """Uses parse_alignment function convert bam file to a dataframe.
-
-    Args:
-     bam: File name of bam file to process.
-
-    Returns:
-     Dataframe with columns:
-     -'slice_name'
-     -'parent_read'
-     -'pe'
-     -'slice'
-     -'mapped'
-     -'multimapped'
-     -'chrom'
-     -'start'
-     -'end'
-     -'coordinates'
-    """
-
-    df_bam = pd.DataFrame(
-        [
-            parse_alignment(aln)
-            for aln in pysam.AlignmentFile(bam, "rb").fetch(until_eof=True)
-        ],
-        columns=[
-            "slice_name",
-            "parent_read",
-            "pe",
-            "slice",
-            "uid",
-            "mapped",
-            "multimapped",
-            "chrom",
-            "start",
-            "end",
-            "coordinates",
-        ],
-    )
-    df_bam.set_index(["slice_name", "chrom", "start"], inplace=True)
-    return df_bam
-
-
-@get_timing(task_name="merging annotations with BAM input")
-def merge_annotations(df, annotations):
-    """Combines annotations with the parsed bam file output.
-
-    Uses pandas outer join on the indexes to merge annotations
-    e.g. number of capture probe overlaps.
-
-    Annotation tsv must have the index as the first column and this index
-    must have intersecting keys with the first dataframe's index.
-
-    Args:
-     df: pd.Dataframe to merge with annotations
-     annotations: Filename of .tsv to read and merge with df
-
-    Returns:
-     Merged dataframe
-
-    """
-
-    # Update, now using chrom and start to stop issues with multimapping
-    df_ann = pd.read_csv(
-        annotations, sep="\t", header=0, index_col=["slice_name", "chrom", "start"], low_memory=False,
-    )
-    df_ann = df_ann.drop(columns="end", errors="ignore")
-
-    return (
-        df.join(df_ann, how="inner")
-        .drop(columns=["slice_name.1"], errors="ignore")
-        .assign(restriction_fragment=lambda df: df['restriction_fragment'].replace('.', 0).astype(int))
-        .reset_index()
-        .sort_values(['parent_read', 'slice'])
-    )
 
 
 class SliceFilter:
@@ -859,64 +701,3 @@ class TiledCSliceFilter(SliceFilter):
             .loc[~multicapture_fragments]
             .reset_index()
         )
-
-
-@get_timing(task_name="analysis of bam file")
-def main(
-    input_bam,
-    annotations,
-    output_prefix,
-    stats_output,
-    method="capture",
-    sample_name="",
-    read_type="",
-):
-
-    # Read bam file and merege annotations
-    df_alignment = parse_bam(input_bam)
-    df_alignment = merge_annotations(df_alignment, annotations)
-
-    slice_filters_dict = {
-        "capture": CCSliceFilter,
-        "tri": TriCSliceFilter,
-        "tiled": TiledCSliceFilter,
-    }
-
-    # Initialise SliceFilter with default args
-    print(f"Filtering slices with method: {method}")
-    slice_filter_type = slice_filters_dict[method]
-    slice_filter = slice_filter_type(slices=df_alignment, sample_name=sample_name, read_type=read_type)
-
-    # Filter slices using the slice_filter
-    slice_filter.filter_slices()
-
-    # Save filtering statisics
-    slice_filter.filter_stats.to_csv(f"{stats_output}.slice.stats.csv", index=False)
-    slice_filter.read_stats.to_csv(f"{stats_output}.read.stats.csv", index=False)
-
-    # Save reporter stats
-    slice_filter.cis_or_trans_stats.to_csv(
-        f"{stats_output}.reporter.stats.csv", index=False
-    )
-
-    # Output slices filtered by capture site
-    for capture_site, df_cap in slice_filter.slices.query('capture != "."').groupby(
-        "capture"
-    ):
-
-        # Extract only fragments that appear in the capture dataframe
-        output_slices = slice_filter.slices.loc[
-            lambda df: df["parent_read"].isin(df_cap["parent_read"])
-            ]
-        # Generate a new slice filterer and extract the fragments
-        output_fragments = slice_filter_type(output_slices).fragments
-        
-
-        # Output fragments and slices
-        output_fragments.sort_values("parent_read").to_csv(
-                f"{output_prefix}.{capture_site.strip()}.fragments.tsv.gz", sep="\t", index=False)
-        
-        output_slices.sort_values("slice_name").to_csv(
-                f"{output_prefix}.{capture_site.strip()}.slices.tsv.gz", sep="\t", index=False)
-        
-
